requirements.txt
Flask==2.2.5
celery==5.5.3
redis==4.5.5
dnspython>=2.3.0
requests==2.31.0
python-dotenv==1.0.0
prometheus-client==0.16.0
sentry-sdk==1.22.0

fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
asyncpg==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aioredis==2.0.1
pydantic==1.10.7

pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.1
fakeredis
pytest-mock
cachetools>=5.3.0  # opcional


